+++ src/BIT.Data.Sync/DeltaStoreBase.cs
@@ -62,8 +62,7 @@ namespace BIT.Data.Sync
 
         public abstract Task ResetDeltasStatusAsync(string identity, CancellationToken cancellationToken=default);
 
-    
-
-    
+        public abstract Task<IDelta> GetDeltaAsync(string deltaId, CancellationToken cancellationToken);
+      
     }
 }