+++ src/BIT.Data.Sync/Server/ISyncServerNode.cs
@@ -20,8 +20,14 @@ namespace BIT.Data.Sync.Server
         Task<IEnumerable<IDelta>> GetDeltasFromOtherNodes(string startIndex, string identity, CancellationToken cancellationToken);
         
         Task ProcessDeltasAsync(IEnumerable<IDelta> deltas, CancellationToken cancellationToken);
+        /// <summary>
+        /// The function will fetch all the deltas that are greater than current <paramref name="deltaId"/>
+        /// </summary>
+        /// <param name="deltaId"></param>
+        /// <param name="cancellationToken"></param>
+        /// <returns></returns>
+        Task<IDelta> GetDeltaAsync(string deltaId, CancellationToken cancellationToken);
 
-      
     }